Add support for directories in "exclude"
It is often useful to exclude an entire part of the hierarchy from
the sitemap, and this change lets you do that. The approach is
very basic, and could generate false positives if you want to include
foobar.html but exclude foobar.htm, but for the general case, it works.